refactor query and find command
create errors array

update stats and update commands with refactored query lib

rename query file to matching
